+++ netlogo-gui/src/main/properties/PropertyEditor.scala
@@ -4,12 +4,10 @@ package org.nlogo.properties
 
 abstract class PropertyEditor[T](val accessor: PropertyAccessor[T],
                                  val handlesOwnErrors: Boolean = false)
-         extends javax.swing.JPanel
+         extends PropertyPanel
 {
 
   def changed() // abstract
-
-  val BORDER_PADDING = 5
 
   private val originalValue: T = accessor.get
 
@@ -19,11 +17,4 @@ abstract class PropertyEditor[T](val accessor: PropertyAccessor[T],
   def get: Option[T]
   def set(value: T): Unit
   def apply() { get.foreach(accessor.set) }
-
-  def getConstraints = {
-    val c = new java.awt.GridBagConstraints
-    c.anchor = java.awt.GridBagConstraints.WEST
-    c.insets = new java.awt.Insets(3, 3, 3, 3)
-    c
-  }
 }
